backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-genai==0.4.0
python-dotenv==1.0.1
mediapipe>=0.10.0
opencv-python>=4.8.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
cloudinary>=1.36.0
